+++ cli/src/lib.rs
@@ -1,6 +1,4 @@
-//! CLI for workspace management of anchor programs.
-
-use crate::config::{Config, Program, ProgramWorkspace, WalletPath};
+use crate::config::{AnchorPackage, Config, ConfigOverride, Program, ProgramWorkspace, WithPath};
 use anchor_client::Cluster;
 use anchor_lang::idl::{IdlAccount, IdlInstruction};
 use anchor_lang::{AccountDeserialize, AnchorDeserialize, AnchorSerialize};
@@ -8,9 +6,11 @@ use anchor_syn::idl::Idl;
 use anyhow::{anyhow, Context, Result};
 use clap::Clap;
 use flate2::read::ZlibDecoder;
-use flate2::write::ZlibEncoder;
+use flate2::write::{GzEncoder, ZlibEncoder};
 use flate2::Compression;
 use rand::rngs::OsRng;
+use reqwest::blocking::multipart::{Form, Part};
+use reqwest::blocking::Client;
 use serde::{Deserialize, Serialize};
 use solana_client::rpc_client::RpcClient;
 use solana_client::rpc_config::RpcSendTransactionConfig;
@@ -30,12 +30,12 @@ use std::path::{Path, PathBuf};
 use std::process::{Child, Stdio};
 use std::string::ToString;
 
-mod config;
-mod template;
+pub mod config;
+pub mod template;
 
 // Version of the docker image.
-const VERSION: &str = env!("CARGO_PKG_VERSION");
-const DOCKER_BUILDER_VERSION: &str = VERSION;
+pub const VERSION: &str = env!("CARGO_PKG_VERSION");
+pub const DOCKER_BUILDER_VERSION: &str = VERSION;
 
 #[derive(Debug, Clap)]
 #[clap(version = VERSION)]
@@ -44,16 +44,6 @@ pub struct Opts {
     pub cfg_override: ConfigOverride,
     #[clap(subcommand)]
     pub command: Command,
-}
-
-#[derive(Debug, Clap)]
-pub struct ConfigOverride {
-    /// Cluster override.
-    #[clap(global = true, long = "provider.cluster")]
-    cluster: Option<Cluster>,
-    /// Wallet override.
-    #[clap(global = true, long = "provider.wallet")]
-    wallet: Option<WalletPath>,
 }
 
 #[derive(Debug, Clap)]
@@ -74,6 +64,10 @@ pub enum Command {
         verifiable: bool,
         #[clap(short, long)]
         program_name: Option<String>,
+        /// Version of the Solana toolchain to use. For --verifiable builds
+        /// only.
+        #[clap(short, long)]
+        solana_version: Option<String>,
     },
     /// Verifies the on-chain bytecode matches the locally compiled artifact.
     /// Run this command inside a program subdirectory, i.e., in the dir
@@ -81,6 +75,12 @@ pub enum Command {
     Verify {
         /// The deployed program to compare against.
         program_id: Pubkey,
+        #[clap(short, long)]
+        program_name: Option<String>,
+        /// Version of the Solana toolchain to use. For --verifiable builds
+        /// only.
+        #[clap(short, long)]
+        solana_version: Option<String>,
     },
     /// Runs integration tests against a localnetwork.
     Test {
@@ -149,6 +149,16 @@ pub enum Command {
     Run {
         /// The name of the script to run.
         script: String,
+    },
+    /// Saves an api token from the registry locally.
+    Login {
+        /// API access token.
+        token: String,
+    },
+    /// Publishes a verified build to the Anchor registry.
+    Publish {
+        /// The name of the program to publish.
+        program: String,
     },
 }
 
@@ -230,8 +240,7 @@ pub enum ClusterCommand {
     List,
 }
 
-fn main() -> Result<()> {
-    let opts = Opts::parse();
+pub fn entry(opts: Opts) -> Result<()> {
     match opts.command {
         Command::Init { name, typescript } => init(&opts.cfg_override, name, typescript),
         Command::New { name } => new(&opts.cfg_override, name),
@@ -239,8 +248,21 @@ fn main() -> Result<()> {
             idl,
             verifiable,
             program_name,
-        } => build(&opts.cfg_override, idl, verifiable, program_name),
-        Command::Verify { program_id } => verify(&opts.cfg_override, program_id),
+            solana_version,
+        } => build(
+            &opts.cfg_override,
+            idl,
+            verifiable,
+            program_name,
+            solana_version,
+            None,
+            None,
+        ),
+        Command::Verify {
+            program_id,
+            program_name,
+            solana_version,
+        } => verify(&opts.cfg_override, program_id, program_name, solana_version),
         Command::Deploy { program_name } => deploy(&opts.cfg_override, program_name),
         Command::Upgrade {
             program_id,
@@ -269,6 +291,8 @@ fn main() -> Result<()> {
         Command::Cluster { subcmd } => cluster(subcmd),
         Command::Shell => shell(&opts.cfg_override),
         Command::Run { script } => run(&opts.cfg_override, script),
+        Command::Login { token } => login(&opts.cfg_override, token),
+        Command::Publish { program } => publish(&opts.cfg_override, program),
     }
 }
 
@@ -340,8 +364,8 @@ fn init(cfg_override: &ConfigOverride, name: String, typescript: bool) -> Result
 
 // Creates a new program crate in the `programs/<name>` directory.
 fn new(cfg_override: &ConfigOverride, name: String) -> Result<()> {
-    with_workspace(cfg_override, |_cfg, path, _cargo| {
-        match path.parent() {
+    with_workspace(cfg_override, |cfg, _cargo| {
+        match cfg.path().parent() {
             None => {
                 println!("Unable to make new program");
             }
@@ -368,26 +392,54 @@ fn new_program(name: &str) -> Result<()> {
     Ok(())
 }
 
-fn build(
+pub fn build(
     cfg_override: &ConfigOverride,
     idl: Option<String>,
     verifiable: bool,
     program_name: Option<String>,
+    solana_version: Option<String>,
+    stdout: Option<File>, // Used for the package registry server.
+    stderr: Option<File>, // Used for the package registry server.
 ) -> Result<()> {
-    let (cfg, path, cargo) = Config::discover(cfg_override)?.expect("Not in workspace.");
+    // Change directories to the given `program_name`, if given.
+    let (cfg, _cargo) = Config::discover(cfg_override)?.expect("Not in workspace.");
 
-    if let Some(program_name) = program_name {
+    let mut did_find_program = false;
+    if let Some(program_name) = program_name.as_ref() {
         for program in cfg.read_all_programs()? {
-            let p = program.path.file_name().unwrap().to_str().unwrap();
-            if program_name.as_str() == p {
-                std::env::set_current_dir(&program.path)?;
+            let cargo_toml = program.path.join("Cargo.toml");
+            if !cargo_toml.exists() {
+                return Err(anyhow!(
+                    "Did not find Cargo.toml at the path: {}",
+                    program.path.display()
+                ));
+            }
+            let p_lib_name = config::extract_lib_name(&cargo_toml)?;
+            if program_name.as_str() == p_lib_name {
+                let program_path = cfg
+                    .path()
+                    .parent()
+                    .unwrap()
+                    .canonicalize()?
+                    .join(program.path);
+                std::env::set_current_dir(&program_path)?;
+                did_find_program = true;
+                break;
             }
         }
     }
+    if !did_find_program && program_name.is_some() {
+        return Err(anyhow!(
+            "{} is not part of the workspace",
+            program_name.as_ref().unwrap()
+        ));
+    }
+
+    let (cfg, cargo) = Config::discover(cfg_override)?.expect("Not in workspace.");
     let idl_out = match idl {
         Some(idl) => Some(PathBuf::from(idl)),
         None => {
-            let cfg_parent = match path.parent() {
+            let cfg_parent = match cfg.path().parent() {
                 None => return Err(anyhow!("Invalid Anchor.toml")),
                 Some(parent) => parent,
             };
@@ -395,9 +447,23 @@ fn build(
             Some(cfg_parent.join("target/idl"))
         }
     };
+
+    let solana_version = match solana_version.is_some() {
+        true => solana_version,
+        false => cfg.solana_version.clone(),
+    };
+
     match cargo {
-        None => build_all(&cfg, path, idl_out, verifiable)?,
-        Some(ct) => build_cwd(path.as_path(), ct, idl_out, verifiable)?,
+        None => build_all(&cfg, cfg.path(), idl_out, verifiable, solana_version)?,
+        Some(ct) => build_cwd(
+            &cfg,
+            ct,
+            idl_out,
+            verifiable,
+            solana_version,
+            stdout,
+            stderr,
+        )?,
     };
 
     set_workspace_dir_or_exit();
@@ -406,21 +472,25 @@ fn build(
 }
 
 fn build_all(
-    cfg: &Config,
-    cfg_path: PathBuf,
+    cfg: &WithPath<Config>,
+    cfg_path: &Path,
     idl_out: Option<PathBuf>,
     verifiable: bool,
+    solana_version: Option<String>,
 ) -> Result<()> {
     let cur_dir = std::env::current_dir()?;
     let r = match cfg_path.parent() {
         None => Err(anyhow!("Invalid Anchor.toml at {}", cfg_path.display())),
-        Some(parent) => {
-            for p in cfg.get_program_list(parent.join("programs"))? {
+        Some(_parent) => {
+            for p in cfg.get_program_list()? {
                 build_cwd(
-                    cfg_path.as_path(),
+                    cfg,
                     p.join("Cargo.toml"),
                     idl_out.clone(),
                     verifiable,
+                    solana_version.clone(),
+                    None,
+                    None,
                 )?;
             }
             Ok(())
@@ -432,10 +502,13 @@ fn build_all(
 
 // Runs the build command outside of a workspace.
 fn build_cwd(
-    cfg_path: &Path,
+    cfg: &WithPath<Config>,
     cargo_toml: PathBuf,
     idl_out: Option<PathBuf>,
     verifiable: bool,
+    solana_version: Option<String>,
+    stdout: Option<File>,
+    stderr: Option<File>,
 ) -> Result<()> {
     match cargo_toml.parent() {
         None => return Err(anyhow!("Unable to find parent")),
@@ -443,51 +516,207 @@ fn build_cwd(
     };
     match verifiable {
         false => _build_cwd(idl_out),
-        true => build_cwd_verifiable(cfg_path.parent().unwrap()),
+        true => build_cwd_verifiable(cfg, cargo_toml, solana_version, stdout, stderr),
     }
 }
 
 // Builds an anchor program in a docker image and copies the build artifacts
 // into the `target/` directory.
-fn build_cwd_verifiable(workspace_dir: &Path) -> Result<()> {
-    // Docker vars.
-    let container_name = "anchor-program";
-    let image_name = format!("projectserum/build:v{}", DOCKER_BUILDER_VERSION);
-    let volume_mount = format!(
-        "{}:/workdir",
-        workspace_dir.canonicalize()?.display().to_string()
-    );
-
+fn build_cwd_verifiable(
+    cfg: &WithPath<Config>,
+    cargo_toml: PathBuf,
+    solana_version: Option<String>,
+    stdout: Option<File>,
+    stderr: Option<File>,
+) -> Result<()> {
     // Create output dirs.
+    let workspace_dir = cfg.path().parent().unwrap().canonicalize()?;
     fs::create_dir_all(workspace_dir.join("target/deploy"))?;
     fs::create_dir_all(workspace_dir.join("target/idl"))?;
 
-    // Build the program in docker.
+    let container_name = "anchor-program";
+
+    // Build the binary in docker.
+    let result = docker_build(
+        cfg,
+        container_name,
+        cargo_toml,
+        solana_version,
+        stdout,
+        stderr,
+    );
+
+    // Wipe the generated docker-target dir.
+    println!("Cleaning up the docker target directory");
     let exit = std::process::Command::new("docker")
         .args(&[
-            "run",
-            "--name",
+            "exec",
             container_name,
-            "-v",
-            &volume_mount,
-            &image_name,
-            "anchor",
-            "build",
+            "rm",
+            "-rf",
+            "/workdir/docker-target",
         ])
+        .stdout(Stdio::inherit())
+        .stderr(Stdio::inherit())
+        .output()
+        .map_err(|e| anyhow::format_err!("Docker rm docker-target failed: {}", e.to_string()))?;
+    if !exit.status.success() {
+        return Err(anyhow!("Failed to build program"));
+    }
+
+    // Remove the docker image.
+    println!("Removing the docker image");
+    let exit = std::process::Command::new("docker")
+        .args(&["rm", "-f", container_name])
         .stdout(Stdio::inherit())
         .stderr(Stdio::inherit())
         .output()
         .map_err(|e| anyhow::format_err!("{}", e.to_string()))?;
     if !exit.status.success() {
-        println!("Error building program");
-        return Ok(());
+        println!("Unable to remove docker container");
+        std::process::exit(exit.status.code().unwrap_or(1));
     }
 
-    let idl = extract_idl("src/lib.rs")?;
+    // Build the idl.
+    if let Ok(Some(idl)) = extract_idl("src/lib.rs") {
+        println!("Extracting the IDL");
+        let out_file = workspace_dir.join(format!("target/idl/{}.json", idl.name));
+        write_idl(&idl, OutFile::File(out_file))?;
+    }
+
+    result
+}
+
+fn docker_build(
+    cfg: &WithPath<Config>,
+    container_name: &str,
+    cargo_toml: PathBuf,
+    solana_version: Option<String>,
+    stdout: Option<File>,
+    stderr: Option<File>,
+) -> Result<()> {
+    let binary_name = config::extract_lib_name(&cargo_toml)?;
+
+    // Docker vars.
+    let image_name = cfg.docker();
+    let volume_mount = format!(
+        "{}:/workdir",
+        cfg.path().parent().unwrap().canonicalize()?.display()
+    );
+    println!("Using image {:?}", image_name);
+
+    // Start the docker image running detached in the background.
+    println!("Run docker image");
+    let exit = std::process::Command::new("docker")
+        .args(&[
+            "run",
+            "-it",
+            "-d",
+            "--name",
+            container_name,
+            "--env",
+            "CARGO_TARGET_DIR=/workdir/docker-target",
+            "-v",
+            &volume_mount,
+            &image_name,
+            "bash",
+        ])
+        .stdout(Stdio::inherit())
+        .stderr(Stdio::inherit())
+        .output()
+        .map_err(|e| anyhow::format_err!("Docker build failed: {}", e.to_string()))?;
+    if !exit.status.success() {
+        return Err(anyhow!("Failed to build program"));
+    }
+
+    // Set the solana version in the container, if given. Otherwise use the
+    // default.
+    if let Some(solana_version) = solana_version {
+        println!("Using solana version: {}", solana_version);
+
+        // Fetch the installer.
+        let exit = std::process::Command::new("docker")
+            .args(&[
+                "exec",
+                container_name,
+                "curl",
+                "-sSfL",
+                &format!("https://release.solana.com/v{0}/install", solana_version,),
+                "-o",
+                "solana_installer.sh",
+            ])
+            .stdout(Stdio::inherit())
+            .stderr(Stdio::inherit())
+            .output()
+            .map_err(|e| anyhow!("Failed to set solana version: {:?}", e))?;
+        if !exit.status.success() {
+            return Err(anyhow!("Failed to set solana version"));
+        }
+
+        // Run the installer.
+        let exit = std::process::Command::new("docker")
+            .args(&["exec", container_name, "sh", "solana_installer.sh"])
+            .stdout(Stdio::inherit())
+            .stderr(Stdio::inherit())
+            .output()
+            .map_err(|e| anyhow!("Failed to set solana version: {:?}", e))?;
+        if !exit.status.success() {
+            return Err(anyhow!("Failed to set solana version"));
+        }
+    }
+
+    let manifest_path = pathdiff::diff_paths(
+        cargo_toml.canonicalize()?,
+        cfg.path().parent().unwrap().canonicalize()?,
+    )
+    .ok_or_else(|| anyhow!("Unable to diff paths"))?;
+    println!(
+        "Building {} manifest: {:?}",
+        binary_name,
+        manifest_path.display().to_string()
+    );
+
+    // Execute the build.
+    let exit = std::process::Command::new("docker")
+        .args(&[
+            "exec",
+            container_name,
+            "cargo",
+            "build-bpf",
+            "--manifest-path",
+            &manifest_path.display().to_string(),
+        ])
+        .stdout(match stdout {
+            None => Stdio::inherit(),
+            Some(f) => f.into(),
+        })
+        .stderr(match stderr {
+            None => Stdio::inherit(),
+            Some(f) => f.into(),
+        })
+        .output()
+        .map_err(|e| anyhow::format_err!("Docker build failed: {}", e.to_string()))?;
+    if !exit.status.success() {
+        return Err(anyhow!("Failed to build program"));
+    }
 
     // Copy the binary out of the docker image.
-    let out_file = format!("../../target/deploy/{}.so", idl.name);
-    let bin_artifact = format!("{}:/workdir/target/deploy/{}.so", container_name, idl.name);
+    println!("Copying out the build artifacts");
+    let out_file = cfg
+        .path()
+        .parent()
+        .unwrap()
+        .canonicalize()?
+        .join(format!("target/deploy/{}.so", binary_name))
+        .display()
+        .to_string();
+
+    // This requires the target directory of any built program to be located at
+    // the root of the workspace.
+    let bin_artifact = format!(
+        "{}:/workdir/docker-target/deploy/{}.so",
+        container_name, binary_name
+    );
     let exit = std::process::Command::new("docker")
         .args(&["cp", &bin_artifact, &out_file])
         .stdout(Stdio::inherit())
@@ -495,33 +724,12 @@ fn build_cwd_verifiable(workspace_dir: &Path) -> Result<()> {
         .output()
         .map_err(|e| anyhow::format_err!("{}", e.to_string()))?;
     if !exit.status.success() {
-        return Ok(());
+        return Err(anyhow!(
+            "Failed to copy binary out of docker. Is the target directory set correctly?"
+        ));
     }
 
-    // Copy the idl out of the docker image.
-    let out_file = format!("../../target/idl/{}.json", idl.name);
-    let idl_artifact = format!("{}:/workdir/target/idl/{}.json", container_name, idl.name);
-    let exit = std::process::Command::new("docker")
-        .args(&["cp", &idl_artifact, &out_file])
-        .stdout(Stdio::inherit())
-        .stderr(Stdio::inherit())
-        .output()
-        .map_err(|e| anyhow::format_err!("{}", e.to_string()))?;
-    if !exit.status.success() {
-        return Ok(());
-    }
-
-    // Remove the docker image.
-    let exit = std::process::Command::new("docker")
-        .args(&["rm", container_name])
-        .stdout(Stdio::inherit())
-        .stderr(Stdio::inherit())
-        .output()
-        .map_err(|e| anyhow::format_err!("{}", e.to_string()))?;
-    if !exit.status.success() {
-        std::process::exit(exit.status.code().unwrap_or(1));
-    }
-
+    // Done.
     Ok(())
 }
 
@@ -537,41 +745,115 @@ fn _build_cwd(idl_out: Option<PathBuf>) -> Result<()> {
     }
 
     // Always assume idl is located ar src/lib.rs.
-    let idl = extract_idl("src/lib.rs")?;
+    if let Some(idl) = extract_idl("src/lib.rs")? {
+        let out = match idl_out {
+            None => PathBuf::from(".").join(&idl.name).with_extension("json"),
+            Some(o) => PathBuf::from(&o.join(&idl.name).with_extension("json")),
+        };
 
-    let out = match idl_out {
-        None => PathBuf::from(".").join(&idl.name).with_extension("json"),
-        Some(o) => PathBuf::from(&o.join(&idl.name).with_extension("json")),
-    };
+        write_idl(&idl, OutFile::File(out))?;
+    }
 
-    write_idl(&idl, OutFile::File(out))
+    Ok(())
 }
 
-fn verify(cfg_override: &ConfigOverride, program_id: Pubkey) -> Result<()> {
-    let (cfg, _path, cargo) = Config::discover(cfg_override)?.expect("Not in workspace.");
-    let cargo = cargo.ok_or_else(|| anyhow!("Must be inside program subdirectory."))?;
+fn verify(
+    cfg_override: &ConfigOverride,
+    program_id: Pubkey,
+    program_name: Option<String>,
+    solana_version: Option<String>,
+) -> Result<()> {
+    // Change directories to the given `program_name`, if given.
+    let (cfg, _cargo) = Config::discover(cfg_override)?.expect("Not in workspace.");
+    let mut did_find_program = false;
+    if let Some(program_name) = program_name.as_ref() {
+        for program in cfg.read_all_programs()? {
+            let cargo_toml = program.path.join("Cargo.toml");
+            if !cargo_toml.exists() {
+                return Err(anyhow!(
+                    "Did not find Cargo.toml at the path: {}",
+                    program.path.display()
+                ));
+            }
+            let p_lib_name = config::extract_lib_name(&cargo_toml)?;
+            if program_name.as_str() == p_lib_name {
+                let program_path = cfg
+                    .path()
+                    .parent()
+                    .unwrap()
+                    .canonicalize()?
+                    .join(program.path);
+                std::env::set_current_dir(&program_path)?;
+                did_find_program = true;
+                break;
+            }
+        }
+    }
+    if !did_find_program && program_name.is_some() {
+        return Err(anyhow!(
+            "{} is not part of the workspace",
+            program_name.as_ref().unwrap()
+        ));
+    }
+
+    // Proceed with the command.
+    let (cfg, cargo) = Config::discover(cfg_override)?.expect("Not in workspace.");
+    let cargo = cargo.ok_or_else(|| {
+        anyhow!("Must be inside program subdirectory if no program name is given.")
+    })?;
     let program_dir = cargo.parent().unwrap();
 
     // Build the program we want to verify.
     let cur_dir = std::env::current_dir()?;
-    build(cfg_override, None, true, None)?;
+    build(
+        cfg_override,
+        None,
+        true,
+        None,
+        match solana_version.is_some() {
+            true => solana_version,
+            false => cfg.solana_version.clone(),
+        },
+        None,
+        None,
+    )?;
     std::env::set_current_dir(&cur_dir)?;
 
-    let local_idl = extract_idl("src/lib.rs")?;
-
     // Verify binary.
-    let bin_path = program_dir
-        .join("../../target/deploy/")
-        .join(format!("{}.so", local_idl.name));
-    let is_buffer = verify_bin(program_id, &bin_path, cfg.provider.cluster.url())?;
+    let binary_name = {
+        let cargo_toml = cargo_toml::Manifest::from_path(&cargo)?;
+        match cargo_toml.lib {
+            None => {
+                cargo_toml
+                    .package
+                    .ok_or_else(|| anyhow!("Package section not provided"))?
+                    .name
+            }
+            Some(lib) => lib.name.ok_or_else(|| anyhow!("Name not provided"))?,
+        }
+    };
+    let bin_path = cfg
+        .path()
+        .parent()
+        .ok_or_else(|| anyhow!("Unable to find workspace root"))?
+        .join("target/deploy/")
+        .join(format!("{}.so", binary_name));
+
+    let bin_ver = verify_bin(program_id, &bin_path, cfg.provider.cluster.url())?;
+    if !bin_ver.is_verified {
+        println!("Error: Binaries don't match");
+        std::process::exit(1);
+    }
 
     // Verify IDL (only if it's not a buffer account).
-    if !is_buffer {
-        std::env::set_current_dir(program_dir)?;
-        let deployed_idl = fetch_idl(cfg_override, program_id)?;
-        if local_idl != deployed_idl {
-            println!("Error: IDLs don't match");
-            std::process::exit(1);
+    if let Some(local_idl) = extract_idl("src/lib.rs")? {
+        if bin_ver.state != BinVerificationState::Buffer {
+            std::env::set_current_dir(program_dir)?;
+            let deployed_idl = fetch_idl(cfg_override, program_id)?;
+            if local_idl != deployed_idl {
+                println!("Error: IDLs don't match");
+                std::process::exit(1);
+            }
         }
     }
 
@@ -580,11 +862,11 @@ fn verify(cfg_override: &ConfigOverride, program_id: Pubkey) -> Result<()> {
     Ok(())
 }
 
-fn verify_bin(program_id: Pubkey, bin_path: &Path, cluster: &str) -> Result<bool> {
+pub fn verify_bin(program_id: Pubkey, bin_path: &Path, cluster: &str) -> Result<BinVerification> {
     let client = RpcClient::new(cluster.to_string());
 
     // Get the deployed build artifacts.
-    let (deployed_bin, is_buffer) = {
+    let (deployed_bin, state) = {
         let account = client
             .get_account_with_commitment(&program_id, CommitmentConfig::default())?
             .value
@@ -592,18 +874,35 @@ fn verify_bin(program_id: Pubkey, bin_path: &Path, cluster: &str) -> Result<bool
         match account.state()? {
             UpgradeableLoaderState::Program {
                 programdata_address,
-            } => (
-                client
+            } => {
+                let account = client
                     .get_account_with_commitment(&programdata_address, CommitmentConfig::default())?
                     .value
-                    .map_or(Err(anyhow!("Account not found")), Ok)?
-                    .data[UpgradeableLoaderState::programdata_data_offset().unwrap_or(0)..]
-                    .to_vec(),
-                false,
-            ),
+                    .map_or(Err(anyhow!("Account not found")), Ok)?;
+                let bin = account.data
+                    [UpgradeableLoaderState::programdata_data_offset().unwrap_or(0)..]
+                    .to_vec();
+
+                if let UpgradeableLoaderState::ProgramData {
+                    slot,
+                    upgrade_authority_address,
+                } = account.state()?
+                {
+                    let state = BinVerificationState::ProgramData {
+                        slot,
+                        upgrade_authority_address,
+                    };
+                    (bin, state)
+                } else {
+                    return Err(anyhow!("Expected program data"));
+                }
+            }
             UpgradeableLoaderState::Buffer { .. } => {
                 let offset = UpgradeableLoaderState::buffer_data_offset().unwrap_or(0);
-                (account.data[offset..].to_vec(), true)
+                (
+                    account.data[offset..].to_vec(),
+                    BinVerificationState::Buffer,
+                )
             }
             _ => return Err(anyhow!("Invalid program id")),
         }
@@ -622,12 +921,24 @@ fn verify_bin(program_id: Pubkey, bin_path: &Path, cluster: &str) -> Result<bool
     }
 
     // Finally, check the bytes.
-    if local_bin != deployed_bin {
-        println!("Error: Binaries don't match");
-        std::process::exit(1);
-    }
+    let is_verified = local_bin == deployed_bin;
 
-    Ok(is_buffer)
+    Ok(BinVerification { state, is_verified })
+}
+
+#[derive(PartialEq)]
+pub struct BinVerification {
+    pub state: BinVerificationState,
+    pub is_verified: bool,
+}
+
+#[derive(PartialEq)]
+pub enum BinVerificationState {
+    Buffer,
+    ProgramData {
+        slot: u64,
+        upgrade_authority_address: Option<Pubkey>,
+    },
 }
 
 // Fetches an IDL for the given program_id.
@@ -660,7 +971,7 @@ fn fetch_idl(cfg_override: &ConfigOverride, idl_addr: Pubkey) -> Result<Idl> {
     serde_json::from_slice(&s[..]).map_err(Into::into)
 }
 
-fn extract_idl(file: &str) -> Result<Idl> {
+fn extract_idl(file: &str) -> Result<Option<Idl>> {
     let file = shellexpand::tilde(file);
     anchor_syn::idl::file::parse(&*file)
 }
@@ -695,7 +1006,7 @@ fn idl(cfg_override: &ConfigOverride, subcmd: IdlCommand) -> Result<()> {
 }
 
 fn idl_init(cfg_override: &ConfigOverride, program_id: Pubkey, idl_filepath: String) -> Result<()> {
-    with_workspace(cfg_override, |cfg, _path, _cargo| {
+    with_workspace(cfg_override, |cfg, _cargo| {
         let keypair = cfg.provider.wallet.to_string();
 
         let bytes = std::fs::read(idl_filepath)?;
@@ -713,7 +1024,7 @@ fn idl_write_buffer(
     program_id: Pubkey,
     idl_filepath: String,
 ) -> Result<Pubkey> {
-    with_workspace(cfg_override, |cfg, _path, _cargo| {
+    with_workspace(cfg_override, |cfg, _cargo| {
         let keypair = cfg.provider.wallet.to_string();
 
         let bytes = std::fs::read(idl_filepath)?;
@@ -729,7 +1040,7 @@ fn idl_write_buffer(
 }
 
 fn idl_set_buffer(cfg_override: &ConfigOverride, program_id: Pubkey, buffer: Pubkey) -> Result<()> {
-    with_workspace(cfg_override, |cfg, _path, _cargo| {
+    with_workspace(cfg_override, |cfg, _cargo| {
         let keypair = solana_sdk::signature::read_keypair_file(&cfg.provider.wallet.to_string())
             .map_err(|_| anyhow!("Unable to read keypair file"))?;
         let client = RpcClient::new(cfg.provider.cluster.url().to_string());
@@ -783,7 +1094,7 @@ fn idl_upgrade(
 }
 
 fn idl_authority(cfg_override: &ConfigOverride, program_id: Pubkey) -> Result<()> {
-    with_workspace(cfg_override, |cfg, _path, _cargo| {
+    with_workspace(cfg_override, |cfg, _cargo| {
         let client = RpcClient::new(cfg.provider.cluster.url().to_string());
         let idl_address = {
             let account = client
@@ -813,7 +1124,7 @@ fn idl_set_authority(
     address: Option<Pubkey>,
     new_authority: Pubkey,
 ) -> Result<()> {
-    with_workspace(cfg_override, |cfg, _path, _cargo| {
+    with_workspace(cfg_override, |cfg, _cargo| {
         // Misc.
         let idl_address = match address {
             None => IdlAccount::address(&program_id),
@@ -945,7 +1256,7 @@ fn idl_write(cfg: &Config, program_id: &Pubkey, idl: &Idl, idl_address: Pubkey) 
 }
 
 fn idl_parse(file: String, out: Option<String>) -> Result<()> {
-    let idl = extract_idl(&file)?;
+    let idl = extract_idl(&file)?.ok_or_else(|| anyhow!("IDL not parsed"))?;
     let out = match out {
         None => OutFile::Stdout,
         Some(out) => OutFile::File(PathBuf::from(out)),
@@ -984,10 +1295,10 @@ fn test(
     skip_build: bool,
     extra_args: Vec<String>,
 ) -> Result<()> {
-    with_workspace(cfg_override, |cfg, _path, _cargo| {
+    with_workspace(cfg_override, |cfg, _cargo| {
         // Build if needed.
         if !skip_build {
-            build(cfg_override, None, false, None)?;
+            build(cfg_override, None, false, None, None, None, None)?;
         }
 
         // Run the deploy against the cluster in two cases:
@@ -1064,15 +1375,15 @@ fn test(
 
 // Returns the solana-test-validator flags to embed the workspace programs
 // in the genesis block. This allows us to run tests without every deploying.
-fn genesis_flags(cfg: &Config) -> Result<Vec<String>> {
-    let clusters = cfg.clusters.get(&Cluster::Localnet);
+fn genesis_flags(cfg: &WithPath<Config>) -> Result<Vec<String>> {
+    let programs = cfg.programs.get(&Cluster::Localnet);
 
     let mut flags = Vec::new();
     for mut program in cfg.read_all_programs()? {
         let binary_path = program.binary_path().display().to_string();
 
-        let address = clusters
-            .and_then(|m| m.get(&program.idl.name))
+        let address = programs
+            .and_then(|m| m.get(&program.lib_name))
             .map(|deployment| deployment.address.to_string())
             .unwrap_or_else(|| {
                 let kp = Keypair::generate(&mut OsRng);
@@ -1083,14 +1394,16 @@ fn genesis_flags(cfg: &Config) -> Result<Vec<String>> {
         flags.push(address.clone());
         flags.push(binary_path);
 
-        // Add program address to the IDL.
-        program.idl.metadata = Some(serde_json::to_value(IdlTestMetadata { address })?);
+        if let Some(mut idl) = program.idl.as_mut() {
+            // Add program address to the IDL.
+            idl.metadata = Some(serde_json::to_value(IdlTestMetadata { address })?);
 
-        // Persist it.
-        let idl_out = PathBuf::from("target/idl")
-            .join(&program.idl.name)
-            .with_extension("json");
-        write_idl(&program.idl, OutFile::File(idl_out))?;
+            // Persist it.
+            let idl_out = PathBuf::from("target/idl")
+                .join(&idl.name)
+                .with_extension("json");
+            write_idl(idl, OutFile::File(idl_out))?;
+        }
     }
     if let Some(test) = cfg.test.as_ref() {
         for entry in &test.genesis {
@@ -1102,7 +1415,7 @@ fn genesis_flags(cfg: &Config) -> Result<Vec<String>> {
     Ok(flags)
 }
 
-fn stream_logs(config: &Config) -> Result<Vec<std::process::Child>> {
+fn stream_logs(config: &WithPath<Config>) -> Result<Vec<std::process::Child>> {
     let program_logs_dir = ".anchor/program-logs";
     if Path::new(program_logs_dir).exists() {
         std::fs::remove_dir_all(program_logs_dir)?;
@@ -1121,7 +1434,7 @@ fn stream_logs(config: &Config) -> Result<Vec<std::process::Child>> {
 
         let log_file = File::create(format!(
             "{}/{}.{}.log",
-            program_logs_dir, metadata.address, program.idl.name
+            program_logs_dir, metadata.address, program.lib_name,
         ))?;
         let stdio = std::process::Stdio::from(log_file);
         let child = std::process::Command::new("solana")
@@ -1195,7 +1508,7 @@ fn _deploy(
     cfg_override: &ConfigOverride,
     program_str: Option<String>,
 ) -> Result<Vec<(Pubkey, Program)>> {
-    with_workspace(cfg_override, |cfg, _path, _cargo| {
+    with_workspace(cfg_override, |cfg, _cargo| {
         let url = cfg.provider.cluster.url().to_string();
         let keypair = cfg.provider.wallet.to_string();
 
@@ -1246,16 +1559,18 @@ fn _deploy(
                 std::process::exit(exit.status.code().unwrap_or(1));
             }
 
-            // Add program address to the IDL.
-            program.idl.metadata = Some(serde_json::to_value(IdlTestMetadata {
-                address: program_kp.pubkey().to_string(),
-            })?);
+            if let Some(mut idl) = program.idl.as_mut() {
+                // Add program address to the IDL.
+                idl.metadata = Some(serde_json::to_value(IdlTestMetadata {
+                    address: program_kp.pubkey().to_string(),
+                })?);
 
-            // Persist it.
-            let idl_out = PathBuf::from("target/idl")
-                .join(&program.idl.name)
-                .with_extension("json");
-            write_idl(&program.idl, OutFile::File(idl_out))?;
+                // Persist it.
+                let idl_out = PathBuf::from("target/idl")
+                    .join(&idl.name)
+                    .with_extension("json");
+                write_idl(idl, OutFile::File(idl_out))?;
+            }
 
             programs.push((program_kp.pubkey(), program))
         }
@@ -1274,7 +1589,7 @@ fn upgrade(
     let path: PathBuf = program_filepath.parse().unwrap();
     let program_filepath = path.canonicalize()?.display().to_string();
 
-    with_workspace(cfg_override, |cfg, _path, _cargo| {
+    with_workspace(cfg_override, |cfg, _cargo| {
         let exit = std::process::Command::new("solana")
             .arg("program")
             .arg("deploy")
@@ -1303,17 +1618,27 @@ fn launch(
     program_name: Option<String>,
 ) -> Result<()> {
     // Build and deploy.
-    build(cfg_override, None, verifiable, program_name.clone())?;
+    build(
+        cfg_override,
+        None,
+        verifiable,
+        program_name.clone(),
+        None,
+        None,
+        None,
+    )?;
     let programs = _deploy(cfg_override, program_name)?;
 
-    with_workspace(cfg_override, |cfg, _path, _cargo| {
+    with_workspace(cfg_override, |cfg, _cargo| {
         let keypair = cfg.provider.wallet.to_string();
 
         // Add metadata to all IDLs.
         for (address, program) in programs {
-            // Store the IDL on chain.
-            let idl_address = create_idl_account(cfg, &keypair, &address, &program.idl)?;
-            println!("IDL account created: {}", idl_address.to_string());
+            if let Some(idl) = program.idl.as_ref() {
+                // Store the IDL on chain.
+                let idl_address = create_idl_account(cfg, &keypair, &address, idl)?;
+                println!("IDL account created: {}", idl_address.to_string());
+            }
         }
 
         // Run migration script.
@@ -1331,7 +1656,7 @@ fn launch(
 fn clear_program_keys(cfg_override: &ConfigOverride) -> Result<()> {
     let config = Config::discover(cfg_override)
         .unwrap_or_default()
-        .unwrap_or_default()
+        .unwrap()
         .0;
 
     for program in config.read_all_programs()? {
@@ -1476,7 +1801,7 @@ fn serialize_idl_ix(ix_inner: anchor_lang::idl::IdlInstruction) -> Result<Vec<u8
 }
 
 fn migrate(cfg_override: &ConfigOverride) -> Result<()> {
-    with_workspace(cfg_override, |cfg, _path, _cargo| {
+    with_workspace(cfg_override, |cfg, _cargo| {
         println!("Running migration deploy script");
 
         let url = cfg.provider.cluster.url().to_string();
@@ -1538,8 +1863,8 @@ fn set_workspace_dir_or_exit() {
             println!("Not in anchor workspace.");
             std::process::exit(1);
         }
-        Some((_cfg, cfg_path, _inside_cargo)) => {
-            match cfg_path.parent() {
+        Some((cfg, _inside_cargo)) => {
+            match cfg.path().parent() {
                 None => {
                     println!("Unable to make new program");
                 }
@@ -1587,15 +1912,22 @@ fn cluster(_cmd: ClusterCommand) -> Result<()> {
 }
 
 fn shell(cfg_override: &ConfigOverride) -> Result<()> {
-    with_workspace(cfg_override, |cfg, _path, _cargo| {
+    with_workspace(cfg_override, |cfg, _cargo| {
         let programs = {
+            // Create idl map from all workspace programs.
             let mut idls: HashMap<String, Idl> = cfg
                 .read_all_programs()?
                 .iter()
-                .map(|program| (program.idl.name.clone(), program.idl.clone()))
+                .filter(|program| program.idl.is_some())
+                .map(|program| {
+                    (
+                        program.idl.as_ref().unwrap().name.clone(),
+                        program.idl.clone().unwrap(),
+                    )
+                })
                 .collect();
             // Insert all manually specified idls into the idl map.
-            if let Some(programs) = cfg.clusters.get(&cfg.provider.cluster) {
+            if let Some(programs) = cfg.programs.get(&cfg.provider.cluster) {
                 let _ = programs
                     .iter()
                     .map(|(name, pd)| {
@@ -1608,20 +1940,20 @@ fn shell(cfg_override: &ConfigOverride) -> Result<()> {
                     })
                     .collect::<Vec<_>>();
             }
-            match cfg.clusters.get(&cfg.provider.cluster) {
+            // Finalize program list with all programs with IDLs.
+            match cfg.programs.get(&cfg.provider.cluster) {
                 None => Vec::new(),
                 Some(programs) => programs
                     .iter()
-                    .map(|(name, program_deployment)| ProgramWorkspace {
-                        name: name.to_string(),
-                        program_id: program_deployment.address,
-                        idl: match idls.get(name) {
-                            None => {
-                                println!("Unable to find IDL for {}", name);
-                                std::process::exit(1);
-                            }
-                            Some(idl) => idl.clone(),
-                        },
+                    .filter_map(|(name, program_deployment)| {
+                        Some(ProgramWorkspace {
+                            name: name.to_string(),
+                            program_id: program_deployment.address,
+                            idl: match idls.get(name) {
+                                None => return None,
+                                Some(idl) => idl.clone(),
+                            },
+                        })
                     })
                     .collect::<Vec<ProgramWorkspace>>(),
             }
@@ -1647,7 +1979,7 @@ fn shell(cfg_override: &ConfigOverride) -> Result<()> {
 }
 
 fn run(cfg_override: &ConfigOverride, script: String) -> Result<()> {
-    with_workspace(cfg_override, |cfg, _path, _cargo| {
+    with_workspace(cfg_override, |cfg, _cargo| {
         let script = cfg
             .scripts
             .get(&script)
@@ -1666,6 +1998,134 @@ fn run(cfg_override: &ConfigOverride, script: String) -> Result<()> {
     })
 }
 
+fn login(_cfg_override: &ConfigOverride, token: String) -> Result<()> {
+    let dir = shellexpand::tilde("~/.config/anchor");
+    if !Path::new(&dir.to_string()).exists() {
+        fs::create_dir(dir.to_string())?;
+    }
+
+    std::env::set_current_dir(dir.to_string())?;
+
+    // Freely overwrite the entire file since it's not used for anything else.
+    let mut file = File::create("credentials")?;
+    file.write_all(template::credentials(&token).as_bytes())?;
+    Ok(())
+}
+
+fn publish(cfg_override: &ConfigOverride, program_name: String) -> Result<()> {
+    // Discover the various workspace configs.
+    let (cfg, _cargo_path) = Config::discover(cfg_override)?.expect("Not in workspace.");
+
+    if !Path::new("Cargo.lock").exists() {
+        return Err(anyhow!("Cargo.lock must exist for a verifiable build"));
+    }
+
+    println!("Publishing will make your code public. Are you sure? Enter (yes)/no:");
+
+    let answer = std::io::stdin().lock().lines().next().unwrap().unwrap();
+    if answer != "yes" {
+        println!("Aborting");
+        return Ok(());
+    }
+
+    let anchor_package = AnchorPackage::from(program_name.clone(), &cfg)?;
+    let anchor_package_bytes = serde_json::to_vec(&anchor_package)?;
+
+    // Build the program before sending it to the server.
+    build(
+        cfg_override,
+        None,
+        true,
+        Some(program_name.clone()),
+        cfg.solana_version.clone(),
+        None,
+        None,
+    )?;
+
+    // Set directory to top of the workspace.
+    let workspace_dir = cfg.path().parent().unwrap();
+    std::env::set_current_dir(workspace_dir)?;
+
+    // Create the workspace tarball.
+    let dot_anchor = workspace_dir.join(".anchor");
+    fs::create_dir_all(&dot_anchor)?;
+    let tarball_filename = dot_anchor.join(format!("{}.tar.gz", program_name));
+    let tar_gz = File::create(&tarball_filename)?;
+    let enc = GzEncoder::new(tar_gz, Compression::default());
+    let mut tar = tar::Builder::new(enc);
+
+    // Files that will always be included if they exist.
+    tar.append_path("Anchor.toml")?;
+    tar.append_path("Cargo.lock")?;
+    if Path::new("Cargo.toml").exists() {
+        tar.append_path("Cargo.toml")?;
+    }
+    if Path::new("LICENSE").exists() {
+        tar.append_path("LICENSE")?;
+    }
+    if Path::new("README.md").exists() {
+        tar.append_path("README.md")?;
+    }
+
+    // All workspace programs.
+    for path in cfg.get_program_list()? {
+        let mut relative_path = pathdiff::diff_paths(path, cfg.path().parent().unwrap())
+            .ok_or_else(|| anyhow!("Unable to diff paths"))?;
+
+        // HACK for workspaces wtih single programs. Change this.
+        if relative_path.display().to_string() == *"" {
+            relative_path = "src".into();
+        }
+        tar.append_dir_all(relative_path.clone(), relative_path)?;
+    }
+    tar.into_inner()?;
+
+    // Upload the tarball to the server.
+    let token = registry_api_token(cfg_override)?;
+    let form = Form::new()
+        .part("manifest", Part::bytes(anchor_package_bytes))
+        .part("workspace", {
+            let file = File::open(&tarball_filename)?;
+            Part::reader(file)
+        });
+    let client = Client::new();
+    let resp = client
+        .post(&format!("{}/api/v0/build", cfg.registry.url))
+        .bearer_auth(token)
+        .multipart(form)
+        .send()?;
+
+    if resp.status() == 200 {
+        println!("Build triggered");
+    } else {
+        println!(
+            "{:?}",
+            resp.text().unwrap_or_else(|_| "Server error".to_string())
+        );
+    }
+
+    Ok(())
+}
+
+fn registry_api_token(_cfg_override: &ConfigOverride) -> Result<String> {
+    #[derive(Debug, Deserialize)]
+    struct Registry {
+        token: String,
+    }
+    #[derive(Debug, Deserialize)]
+    struct Credentials {
+        registry: Registry,
+    }
+    let filename = shellexpand::tilde("~/.config/anchor/credentials");
+    let mut file = File::open(filename.to_string())?;
+    let mut contents = String::new();
+    file.read_to_string(&mut contents)?;
+
+    let credentials_toml: Credentials = toml::from_str(&contents)?;
+
+    Ok(credentials_toml.registry.token)
+}
+
 // with_workspace ensures the current working directory is always the top level
 // workspace directory, i.e., where the `Anchor.toml` file is located, before
 // and after the closure invocation.
@@ -1674,17 +2134,17 @@ fn run(cfg_override: &ConfigOverride, script: String) -> Result<()> {
 // to be outside the workspace. Doing so will have undefined behavior.
 fn with_workspace<R>(
     cfg_override: &ConfigOverride,
-    f: impl FnOnce(&Config, PathBuf, Option<PathBuf>) -> R,
+    f: impl FnOnce(&WithPath<Config>, Option<PathBuf>) -> R,
 ) -> R {
     set_workspace_dir_or_exit();
 
     clear_program_keys(cfg_override).unwrap();
 
-    let (cfg, cfg_path, cargo_toml) = Config::discover(cfg_override)
+    let (cfg, cargo_toml) = Config::discover(cfg_override)
         .expect("Previously set the workspace dir")
         .expect("Anchor.toml must always exist");
 
-    let r = f(&cfg, cfg_path, cargo_toml);
+    let r = f(&cfg, cargo_toml);
 
     set_workspace_dir_or_exit();
     clear_program_keys(cfg_override).unwrap();